# schema/openapi.yaml
...
openapi: 3.1.0
info:
  title: slim-m API
  version: 0.5.0
  description: >
    The server surface: liveness, auth with opaque rotating tokens, channel
    management, text messaging with reactions, polls, deletion, and full-text
    search, user profiles and the member list, devices, blocking, report
    intake and triage, invite codes, read state, bundled catch-up sync,
    account deletion, admin-issued password reset, role management and
    channel permission overwrites, presence and typing indicators, and the
    WebSocket event envelope.

    Identity and ordering are separate concerns. A message's identity is a
    client-generated UUIDv7, which also makes sending idempotent; its order is a
    per-channel monotonic `seq` assigned by the server, which is the
    authoritative cursor for pagination and sync.

    Authentication is a bearer access token. Access tokens are short-lived and
    exchanged at `/auth/refresh`, which rotates the refresh token. Opening a
    WebSocket uses a single-use ticket from `/auth/ws-ticket`, because an
    upgrade request cannot reliably carry an Authorization header.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /channels/{channelId}/messages/polls:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    post:
      operationId: sendPollMessage
      tags: [messages]
      summary: Send a message that carries a poll.
      description: >
        A poll is not its own resource: it is a message, so this is
        idempotent by the client-supplied `id` exactly like `sendMessage`,
        and gated on the same VIEW_CHANNEL plus SEND_MESSAGES permissions,
        evaluated in this channel. Options are ordered and fixed at
        creation; a poll needs between 2 and 4 of them. `close_at`, if given,
        is a unix-millisecond deadline after which the server refuses votes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePollMessageRequest"
      responses:
        "200":
          description: The stored message, with its poll attached.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /channels/{channelId}/messages/{messageId}/pin:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
      - name: messageId
        in: path
        required: true
        description: The message's UUID.
        schema:
          type: string
          format: uuid
    put:
      operationId: pinMessage
      tags: [messages]
      summary: Pin a message in its channel.
      description: >
        Idempotent: pinning an already-pinned message leaves the original
        pin's timestamp and pinner in place. Requires MANAGE_MESSAGES,
        evaluated in this channel specifically; holding it in another channel
        is not enough. Live connections are told over the WebSocket
        (`message.pinned`).
      responses:
        "204":
          description: The message was pinned (or already was).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      operationId: unpinMessage
      tags: [messages]
      summary: Unpin a message.
      description: >
        Idempotent: unpinning a message that is not pinned succeeds, since the
        caller's intent already holds either way. Requires MANAGE_MESSAGES in
        this channel. Live connections are told over the WebSocket
        (`message.unpinned`).
      responses:
        "204":
          description: The message was unpinned (or was already not pinned).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /channels/{channelId}/pins:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    get:
      operationId: listPinnedMessages
      tags: [messages]
      summary: List a channel's pinned messages, newest pin first.
      description: >
        Requires only VIEW_CHANNEL. Each entry is the full message, flattened,
        plus when and by whom it was pinned.
      responses:
        "200":
          description: The channel's pinned messages, newest pin first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PinnedMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /channels/{channelId}/pins/count:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    get:
      operationId: getPinnedMessageCount
      tags: [messages]
      summary: How many messages are pinned in a channel.
      description: >
        Backs the channel header's pin count without the client fetching
        every pinned message just to count them. Requires only VIEW_CHANNEL.
      responses:
        "200":
          description: The pin count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinCount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /messages/{messageId}/polls/vote:
    parameters:
      - name: messageId
        in: path
        required: true
        description: The message's UUID.
        schema:
          type: string
          format: uuid
    put:
      operationId: votePoll
      tags: [messages]
      summary: Cast or change the caller's vote on a message's poll.
      description: >
        One vote per user per poll: a second call replaces the first rather
        than adding to it. Requires being able to view the message's channel
        as well as SEND_MESSAGES there - the same permission sending a
        message needs, evaluated in this channel specifically. Refused once
        the poll's `close_at` has passed, checked against the server's own
        clock. Live connections are told over the WebSocket (`poll.voted`)
        with the whole refreshed tally, never who cast which vote.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VotePollRequest"
      responses:
        "204":
          description: The vote was recorded.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An opaque access token from register, login, or refresh.
  parameters:
    ChannelId:
      name: channelId
      in: path
      required: true
      description: The channel's UUID.
      schema:
        type: string
        format: uuid
  responses:
...
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, expired, or revoked credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: >
        The caller lacks the required permission. A channel that does not exist
        is refused identically, so its existence is not observable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The resource does not exist within a scope the caller can see.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request collided with existing state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: The caller is over its rate budget for this traffic class.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: A short, stable, human-readable reason.
...
    Message:
      type: object
      required: [id, channel_id, seq, content, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Client-generated UUIDv7; also the idempotency key.
        channel_id:
          type: string
          format: uuid
        author_id:
          type: [string, "null"]
          format: uuid
          description: Null once the author's account has been deleted.
        author_display_name:
          type: [string, "null"]
          description: >-
            The author's display name, sent with the message so a client does
            not need a lookup per distinct sender. Null once the account is
            anonymized, exactly as author_id is.
        seq:
          type: integer
          format: int64
          description: Per-channel monotonic order key, and the sync cursor.
        content:
          type: string
        created_at:
          type: integer
          format: int64
        edited_at:
          type: [integer, "null"]
          format: int64
        reactions:
          type: array
          description: >-
            Reaction summaries for this message, one entry per distinct emoji,
            with `reacted` set from the calling user's point of view. Always
            present: an empty array means no reactions, never that the server
            omitted them.
          items:
            $ref: "#/components/schemas/ReactionSummary"
        poll:
          oneOf:
            - $ref: "#/components/schemas/Poll"
            - type: "null"
          description: >-
            The poll this message carries, if any. Always present as a key:
            null means this message is not a poll, the same convention
            `reactions` follows for "genuinely none" versus omitted.
        attachments:
          type: array
          description: >-
            Always present, empty when there are none - same convention as
            `reactions`. Unlike reactions, a freshly sent message can carry
            these immediately, since they are uploaded before the send and
            only referenced by it.
          items:
            $ref: "#/components/schemas/Attachment"
    Attachment:
      type: object
      required: [id, filename, content_type, size]
      properties:
        id:
          type: string
          description: >
            Hex sha256 of the stored bytes. Also the path segment
            `getAttachment` fetches it at.
        filename:
          type: string
          description: >
            Sanitized display name. Content-addressed storage means this is a
            property of the bytes, not of any one message attaching them: the
            first upload of a given hash sets it.
        content_type:
          type: string
          description: >
            One of a fixed server-side allowlist, decided by sniffing the
            uploaded bytes - never the client's declared Content-Type or a
            filename extension.
        size:
          type: integer
          format: int64
    PinnedMessage:
      description: >-
        A pinned message: every Message field, flattened, plus when and by
        whom it was pinned.
      allOf:
        - $ref: "#/components/schemas/Message"
        - type: object
          required: [pinned_at]
          properties:
            pinned_at:
              type: integer
              format: int64
              description: When the message was pinned, unix milliseconds.
            pinned_by:
              type: [string, "null"]
              format: uuid
              description: >-
                Who pinned it. Null once that account has been anonymized,
                exactly as author_id is.
    PinCount:
      type: object
      required: [count]
      properties:
        count:
          type: integer
          format: int64
          description: >-
            How many messages are currently pinned in the channel. Cheap to
            read: a single indexed count, never a fetch of every pinned
            message.
...
    ReactionSummary:
      type: object
      required: [emoji, count, reacted]
      properties:
        emoji:
          type: string
          description: The reaction emoji itself. User content, never chrome.
        count:
          type: integer
          format: int64
          description: How many people have reacted with this emoji.
        reacted:
          type: boolean
          description: >-
            Whether the calling user is one of them, so a client can render the
            toggled state without a second request.
    Poll:
      type: object
      required: [question, options, total_votes, closed]
      properties:
        question:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 4
          description: Ordered and fixed at creation; never reordered or edited.
          items:
            $ref: "#/components/schemas/PollOption"
        total_votes:
          type: integer
          format: int64
          description: The sum of every option's count.
        voted_option:
          type: [integer, "null"]
          format: int64
          description: >-
            The 0-based position the calling user picked, or null if they
            have not voted. Per-viewer, exactly like ReactionSummary.reacted:
            never present for anyone but the voter themselves.
        close_at:
          type: [integer, "null"]
          format: int64
          description: Unix milliseconds. Null means the poll never closes.
        closed:
          type: boolean
          description: >-
            Whether the server's own clock already considers this poll past
            close_at. Voting is refused server-side regardless of what a
            client believes this to be.
    PollOption:
      type: object
      required: [position, label, votes]
      properties:
        position:
          type: integer
          format: int64
          description: 0-based, fixed at creation; also the option's identity.
        label:
          type: string
        votes:
          type: integer
          format: int64
          description: >-
            How many people picked this option. Who they are is never
            revealed by this or any other read: only the calling user's own
            voted_option is.
    CreatePollMessageRequest:
      type: object
      required: [id, question, options]
      properties:
        id:
          type: string
          format: uuid
          description: >
            Client-generated UUIDv7. Retrying with the same id returns the
            stored message instead of sending twice.
        content:
          type: string
          maxLength: 4000
          description: >-
            An optional caption riding alongside the poll. May be empty or
            omitted: the poll itself is what the message is for.
        question:
          type: string
          minLength: 1
          maxLength: 300
        options:
          type: array
          minItems: 2
          maxItems: 4
          items:
            type: string
            minLength: 1
            maxLength: 100
        close_at:
          type: integer
          format: int64
          description: >-
            Unix milliseconds. Omitted or null means the poll never closes.
    VotePollRequest:
      type: object
      required: [option]
      properties:
        option:
          type: integer
          format: int64
          description: The 0-based position of the option being voted for.
...